+++ CHANGELOG.md
@@ -17,10 +17,12 @@ configured prefix accordingly.
 
 ### Added
 
+- Added console command `version`
 - Added debug bot commands `-argstr` and `-argint`
 - Added bot commands `-repo` and `-issues`
 - Added command methods `cogs.debug.DebugCog.argstr()` and `cogs.debug.DebugCog.argint()`
 - Added command methods `cogs.general.GeneralCog.repo()` and `cogs.general.GeneralCog.issues()`
+- Added method `console.LydianConsole.version()`
 
 ### Changed